Risolto TUTTO

+++ Dichiarazione_Variabili.h
@@ -8,8 +8,8 @@
 const int colonne=2;
 void piuPresente(float m[][colonne], int righe);
 void valoreMaggiore(float m[][colonne], int righe);
-void inserisciProdotti(float m[][colonne], float righe);
-void stampaProdotti(float m[][colonne], float righe);
+void inserisciProdotti(float m[][colonne], int righe);
+void stampaProdotti(float m[][colonne], int righe);
 void seNonAbilitato();
 void totaleProdotto(float m[][colonne], int righe);
 void valoreMagazzino(float m[][colonne], int righe);
